Separate out into spark341db shim
